Wait for both AppFutures and DFK task state to indicate a task is done, at shutdown
AppFutures and the DFK state separately indicate that a task is complete, but
at not exactly the same time.

The AppFuture completing does not indicate that some end-of-task-completion processing
has happened, such as monitoring. Because of this, waiting for all AppFutures and then
exiting the process does not guarantee that all end-of-task-completion code will
happen.

This PR waits for the DFK to indicate that the DFK task record also indicates that
the task is done.

This is related to issue #1279, which more broadly discusses task completion issues.