Make memo function hashing more consistent
Prior to this PR, inputs for memo hashes are generated in two places:

* in app construction, producing a func_hash by attempting to get the
  function source code

* id_for_memo hashes various things, including arguments and the app name

When the function is wrapped by some pieces of parsl, such as
remote_side_bash_executor or some file staging code, then the function is
replaced by a new function, and the original function is passed as an
additional argument.

This has the effect of id_for_memo hashing also hashing the function, in
addition to the app construction hash, and this time hashing it in a
different way - by using serialisation of the function object.

This hash-by-serialisation method was extremely fragile and would change for
probably undesirable reasons, such as source code changes elsewhere in
the source file.

This PR removes the first explicit func_hash construction, and makes all
hash generation happen with id_for_memo.

It replaces the serialisation-based hashing of functions with the source based
method previously used in the app's generation of func_hash. This now applies
to both the base function, and functions passed as arguments.

This PR does not address:

  * should memoisation hashes include definitions of helper functions
    such as remote_side-bash_executor? (they do now)
  * how should functions be hashed? (some use cases do not want to
    invalidate checkpoints if app definitions are changed)